Train_XcFaceNet/requirements-linux-py38.txt
absl-py==2.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
#contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
grpcio==1.67.0
#h5py==3.12.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Markdown==3.7
#MarkupSafe==3.0.2
#matplotlib==3.9.2
matplotlib
mpmath==1.3.0
#networkx==3.4.2
numpy==1.23.0
opencv-python==4.10.0.84
packaging==24.1
#pillow==11.0.0
pillow
protobuf==5.28.2
#pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-learn
scipy==1.8.0
six==1.16.0
sympy==1.13.3
tensorboard
tensorboard-data-server
threadpoolctl
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
